requirements.txt
yfinance
httpx
motor
pydantic
python-dotenv